updated marks calculation method and added init api to get assessment total marks (#140)

+++ assessment_service/src/main/java/vacademy/io/assessment_service/features/assessment/dto/Questio_type_based_dtos/mcqm/MCQMMarkingDto.java
@@ -15,8 +15,8 @@ public class MCQMMarkingDto {
     public static class DataFields {
         private double totalMark;
         private double negativeMark;
-        private int negativeMarkingPercentage;
-        private int partialMarking;
-        private int partialMarkingPercentage;
+        private double negativeMarkingPercentage;
+        private double partialMarking;
+        private double partialMarkingPercentage;
     }
 }
